backend(routes): added put/delete routes to users, and moved routes/tests to routes folder
