+++ server/api.js
@@ -6,7 +6,7 @@ const api = module.exports = require('express').Router() // eslint-disable-line 
 api
   .get('/heartbeat', (req, res) => res.send({ok: true}))
   .use('/auth', require('./auth'))
-  .use('/users', require('./users'))
+  .use('/users', require('./routes/api.users')) // sokmean: comment for merging - keep this line, get rid of the other ones for users. Moved api.users to routes folder.
 
 // No routes matched? 404.
 api.use((req, res) => res.status(404).end())
